move _display to remove circular import
